@@ -1,37 +1,45 @@
 package com.example.myminiview
 
+import android.annotation.SuppressLint
 import android.os.Build
 import android.os.Bundle
+import android.util.Log
 import android.view.WindowInsets
 import android.view.WindowManager
+import android.webkit.ConsoleMessage
 import android.webkit.PermissionRequest
 import android.webkit.WebChromeClient
 import android.webkit.WebSettings
 import android.webkit.WebViewClient
 import androidx.appcompat.app.AppCompatActivity
-import kotlinx.android.synthetic.main.activity_main.webView
+import com.example.myminiview.databinding.ActivityMainBinding
 
 
 class MainActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityMainBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
 
         // WebViewClient allows you to handle
         // onPageFinished and override Url loading.
-        webView.webViewClient = WebViewClient()
-        webView.settings.mediaPlaybackRequiresUserGesture = false;
+        binding.webView.webViewClient = WebViewClient()
+        binding.webView.settings.mediaPlaybackRequiresUserGesture = false;
 
         // this will load the url of the website
-        webView.loadUrl("https://mini-view.chop.dev/")
+        binding.webView.loadUrl("https://mini-view.web.app/")
 
         // this will enable the javascript settings
-        webView.settings.javaScriptEnabled = true
+        binding.webView.settings.javaScriptEnabled = true
+        binding.webView.settings.domStorageEnabled = true
+        binding.webView.settings.databaseEnabled = true
+        binding.webView.settings.allowFileAccess = true
 
         // if you want to enable zoom feature
-        webView.settings.setSupportZoom(true)
+        binding.webView.settings.setSupportZoom(true)
 
         supportActionBar?.hide()
 
@@ -44,11 +52,17 @@ class MainActivity : AppCompatActivity() {
             )
         }
 
-        webView.webChromeClient = object : WebChromeClient() {
+        binding.webView.webChromeClient = object : WebChromeClient() {
             override fun onPermissionRequest(request: PermissionRequest) {
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
                     request.grant(request.resources)
                 }
+            }
+
+            override fun onConsoleMessage(message: ConsoleMessage): Boolean {
+                Log.i("MyApplication", "${message.message()} -- From line " +
+                        "${message.lineNumber()} of ${message.sourceId()}")
+                return true
             }
         }
     }
@@ -56,8 +70,8 @@ class MainActivity : AppCompatActivity() {
     // if you press Back button this code will work
     override fun onBackPressed() {
         // if your webview can go back it will go back
-        if (webView.canGoBack())
-            webView.goBack()
+        if (binding.webView.canGoBack())
+            binding.webView.goBack()
         // if your webview cannot go back
         // it will exit the application
         else
